backend/requirements.txt
sqlalchemy
pymysql
python-dotenv
fastapi
uvicorn
passlib
python-jose
cryptography
python-multipart
pydantic[email]
aioredis